+++ src/features/cleaner_file_size.rs
@@ -13,7 +13,7 @@ pub fn directory_cleaner_based_on_file_size(
     report_data: &mut ReportData,
 ) -> Result<()> {
     let mut del_count: u32 = 0;
-    let mut del_size: u32 = 0;
+    let mut del_size: u64 = 0;
     for entry in WalkDir::new(directory).into_iter() {
         match entry {
             Ok(dir) => {
@@ -26,7 +26,7 @@ pub fn directory_cleaner_based_on_file_size(
                     if metadata.len() >= size {
                         delete_file(path, dry_run)?;
                         del_count += 1;
-                        del_size += metadata.len() as u32;
+                        del_size += metadata.len();
                     }
                     collect_metrics(report_data, metadata, &path, (del_count, del_size));
                     del_count = 0;
